--- a/ENIGMAsystem/SHELL/Bridges/General/DX11Context.h
+++ b/ENIGMAsystem/SHELL/Bridges/General/DX11Context.h
@@ -61,7 +61,6 @@ extern ID3D11RasterizerState* m_rasterState;
 class ContextManager {
 
 float last_depth;
-int last_stride;
 bool hasdrawn;
 int shapes_d3d_model;
 int shapes_d3d_texture;
@@ -74,7 +73,6 @@ ContextManager() {
 	hasdrawn = false;
 	shapes_d3d_model = -1;
 	shapes_d3d_texture = -1;
-	last_stride = -1;
 	last_depth = 0.0f;
 }
 
@@ -113,15 +111,11 @@ int GetShapesModel() {
 void BeginShapesBatching(int texId) {
 	if (shapes_d3d_model == -1) {
 		shapes_d3d_model = d3d_model_create(true);
-		last_stride = -1;
-	} else if (texId != shapes_d3d_texture || (d3d_model_get_stride(shapes_d3d_model) != last_stride && last_stride != -1)) {
-		last_stride = -1;
+	} else if (texId != shapes_d3d_texture) {
 		if (!hasdrawn) {
 			d3d_model_draw(shapes_d3d_model, shapes_d3d_texture);
 			d3d_model_clear(shapes_d3d_model);
 		}
-	} else {
-		last_stride = d3d_model_get_stride(shapes_d3d_model);
 	}
 	hasdrawn = false;
 	shapes_d3d_texture = texId;
@@ -134,7 +128,6 @@ void EndShapesBatching() {
 	d3d_model_draw(shapes_d3d_model, shapes_d3d_texture);
 	d3d_model_clear(shapes_d3d_model);
 	shapes_d3d_texture = -1;
-	last_stride = -1;
 }
 
 };