--- a/ENIGMAsystem/SHELL/Bridges/General/GL3Context.h
+++ b/ENIGMAsystem/SHELL/Bridges/General/GL3Context.h
@@ -37,7 +37,6 @@ protected:
 public:
 
 float last_depth;
-int last_stride;
 bool hasdrawn;
 int shapes_d3d_model;
 int shapes_d3d_texture;
@@ -47,7 +46,6 @@ ContextManager() {
 	hasdrawn = false;
 	shapes_d3d_model = -1;
 	shapes_d3d_texture = -1;
-	last_stride = -1;
 	last_depth = 0.0f;
 	bound_tex = 0;
 }
@@ -87,15 +85,11 @@ void RestoreState() {
 void BeginShapesBatching(int texId) {
 	if (shapes_d3d_model == -1) {
 		shapes_d3d_model = enigma_user::d3d_model_create(enigma_user::model_stream);
-		last_stride = -1;
-	} else if (texId != shapes_d3d_texture || (enigma_user::d3d_model_get_stride(shapes_d3d_model) != last_stride && last_stride != -1)) {
-		last_stride = -1;
+	} else if (texId != shapes_d3d_texture) {
 		if (!hasdrawn) {
 			enigma_user::d3d_model_draw(shapes_d3d_model, shapes_d3d_texture);
 			enigma_user::d3d_model_clear(shapes_d3d_model);
 		}
-	} else {
-		last_stride = enigma_user::d3d_model_get_stride(shapes_d3d_model);
 	}
 	hasdrawn = false;
 	shapes_d3d_texture = texId;
@@ -108,8 +102,7 @@ void EndShapesBatching() {
 	enigma_user::d3d_model_clear(shapes_d3d_model);
 
 	shapes_d3d_texture = -1;
-	last_stride = -1;
-  last_depth -= 1;
+  	last_depth -= 1;
 }
 
 void BeginScene() {
@@ -180,9 +173,9 @@ void BlendFunc() { //Used when calling blend functions
 
 void ColorFunc() { //Used when calling color functions
 	if (shapes_d3d_model != -1){
-		if (enigma_user::d3d_model_has_color(shapes_d3d_model) == false){
+		//if (enigma_user::d3d_model_has_color(shapes_d3d_model) == false){
 			EndShapesBatching();
-		}
+		//}
 	}
 }