This pull request attempts to leverage the new generic vertex and index buffer abstractions to generically implement a new 3D model class to replace the old duplicate ones. This pull request continues alongside #1238 and will also make newer and alternative graphics systems easier to support, such as OpenGL ES or Direct3D11, without sacrificing performance, expressiveness, or compatibility as I will demonstrate in the proceeding benchmarks.
### Noteworthy changes made
* Every old model class removed and deleted from git
* A new generic model class was created to replace the deleted ones
* Changed `d3d_model_create` to return an int like other resource functions for consistency
* Removed `d3d_model_part_draw` functions because their name is inconsistent (e.g, should be `draw_part` like `draw_background_part`) and because they are not clear enough for a model class which optimizes and combines primitives together, `d3d_model_draw_primitive` with an index to the primitive would be a better function. We also no longer use these functions anywhere, tiles use the vertex and index buffer abstractions directly now which offer range arguments in their submit functions.
* Added a new function `d3d_model_vertex_data` that allows one to specify the model's vertex data in a single function call, but requires a vertex format be provided.
* Updated `GSmodel.h` with comments specifying the preconditions for each group of model specification functions
* Renamed `d3d_model_add_*` functions to remove the `_add_` part of their name to match the original `d3d_model_vertex` function from GM for consistency.
* Added the following new vertex functions:
    * `vertex_format_exists()`: This non-id version of the function will tell us if we are between `vertex_format_begin` and `vertex_format_end` calls. In other words, it tells us if we are still defining a new vertex format.
    * `vertex_format_get_stride()`: Non-id version returns the stride of the vertex format currently being specified.
    * `vertex_format_get_hash()`: Non-id version returns the hash of the vertex format currently being specified.
    * `vertex_format_get_stride(id)`: Returns the stride of the vertex format with the given id. This is basically the number of vertex attributes, it is not the size of a vertex in bytes.
    * `vertex_format_get_hash(id)`: Returns a hash of the vertex format and its type/usage pairs.
    * `vertex_set_format(buffer,format)`: Allows one to change the vertex format of a vertex buffer, which essentially allows the vertex buffer to hold the same data for differently formatted primitives.
* Made `vertex_format_end` set the `enigma::vertexFormat` pointer to 0 so there can be no further accidental mutation of it once it is moved to the vector and assigned a real id for the user.
* Made `vertex_get_number`, `vertex_get_size`, and their index buffer equivalents return the correct value when in the middle of `vertex_begin` and `vertex_end` which previously they would not. This required marking the vertex and index buffers dirty in `vertex_begin` instead of `vertex_end` so that the dirty flag could be used to determine where to grab the number of vertices/indices from.
* Added a non-fatal error message for `vertex_data` that shows in debug mode if the user calls the function but the vertex buffer does not have a valid vertex format, since one is required for the function to correctly decompose the arguments.
* Removed all `d3d_model_has_*` functions because we aren't using them anywhere and we don't need them.
* Changed `graphics_reset_client_state` to unbind the vertex and index buffers (bind the target to 0) in OpenGL1 because it will screw up primitive rendering that uses `glVertexPointer` with a client-side array. This caused issues in Wild Racing GMK because `d3d_draw_floor` was not being drawn and I could not figure out why. 
* Removed `d3d_model_get_stride` and `d3d_model_format` as well as calls to them in the context managers in Bridges because the function no longer makes sense. The model class in this pr has a format for each primitive, hence the entire model doesn't have a consistent stride or format.
* Removed `d3d_model_index` because the new model class doesn't support indexing yet.
* Added caching of both vertex formats and vertex format peers. The peers are non-trivial to construct and we also don't want the user being wasteful with the former either. Caching both independently gave separate performance boosts, so both are necessary. Because of this change, `vertex_format_end` may now return the same id multiple times and never actually return a unique id if the hashes of the vertex format specified are always logically equivalent to another format.
* Fixed a typo in `d3d_model_set_rotation_x` for D3D9, it should have had `gs_scalar` not `double` like its signature in the general header.
* Added a `dynamic` flag to `vertex_freeze` so the user can indicate such usage. By default I set the parameter to false, because freezing the vertex is to be synonymous with making it static (never changing). If the user passes true for dynamic, that is like saying they will update the vertex buffer infrequently (aka GL_DYNAMIC_DRAW). If the user intends to update the vertex buffer every frame, then they should not call `vertex_freeze` just like in GMS (aka this indicates GL_STREAM_DRAW).

### Optimization
This pull request wouldn't be complete if I didn't run some benchmarks. I clearly ran several games I have to check the performance, but the results I am going to share here are specifically from testing the GMS cubes demo.

Note: I used GPU-Z to monitor VRAM usage and I took the difference in VRAM occupation just before and after closing the game.
https://www.techpowerup.com/gpuz/

Some consistent issues that exist on master too that I want to note that this pr will _not_ be fixing:
* OpenGL1 transforms are screwed up
* In all systems the lighting has an inconsistency with GMS (the projection functions should not update the light positions, they should just be updated in the draw call if dirty)
* `display_reset` is broke in all of the working systems

![Cubes Demo in ENIGMA DX9](https://user-images.githubusercontent.com/3212801/44048049-4959491e-9efe-11e8-98d4-7b3c3cc5b170.png)


#### Current Master
| Branch     | Model create time (us) | FPS  | RAM (MB) | VRAM (MB) |
|------------|------------------------|------|----------|-----------|
| GMS v1.4   | 520224                 | 1430 | 175.0    | 53        |
| ENIGMA GL1 | 062532                 | 1533 | 076.4    | 55        |
| ENIGMA GL3 | 078157                 | 1557 | 087.6    | 56        |
| ENIGMA DX9 | 078112                 | 0658 | 093.7    | 38        |

#### This pull request unoptimized
| Branch     | Model create time (us) | FPS | RAM (MB)         |
|------------|------------------------|-----|------------------|
| ENIGMA GL1 | 109387                 | 18  | 093.9            |
| ENIGMA GL3 | 109410                 | 05  | 107.4+ (leaking) |
| ENIGMA DX9 | 109416                 | NA  | 090.8            |

#### Adjacent primitive combining
This optimization is fairly obvious. If you have two primitives of the same type that have the same vertex format adjacent in the model, you can simply treat them as a single primitive and thus make a single draw call. It was during this step that I had to provide a mechanism to hash the vertex formats for quick comparison. 

This optimization is one of the reasons why a function like `d3d_model_part_draw` doesn't make sense, because it gets in the way of optimizing the model. Such functionality can be brought back of course by maintaining a mapping between the user specified ranges and what actually ends up in the model.

I saw some significant speedup in OpenGL1 from this optimization, but less so in Direct3D9 and OpenGL3 due to anomalies in their primitive batching (the fonts etc require a different kind of optimization).

| Branch     | Model create time (us) | FPS  | RAM (MB)         |
|------------|------------------------|------|------------------|
| ENIGMA GL1 | 124989                 | 1404 | 091.3            |
| ENIGMA GL3 | 109381                 | 970  | 301.3+ (leaking) |
| ENIGMA DX9 | 125010                 | 15   | 099.6+ (leaking) |

#### Vertex format caching
This optimization is also fairly obvious, but I actually didn't think it would bring the performance that it did. Originally I was trying to add `vertex_format_delete` as a function but gave up because my logic was too tricky. I decided to implement a cache for all of the user vertex formats using an `std::unordered_map` and the hashing value I already added.

As Rusky, Josh, and I discussed privately, this is better than expecting the user to have to worry about deleting and cleaning up vertex formats, which GMS doesn't. So you can freely call `vertex_format_end` as many times as you like, and you may just end up being given an already created vertex format.

Doing it in the general source is best because it doesn't require that I also add it to each system. Adding it for each peer vertex format would just be redundant, the D3D/OGL backends use the same ids.

| Branch     | Model create time (us) | FPS  | RAM (MB) |
|------------|------------------------|------|----------|
| ENIGMA GL1 | 109385                 | 1405 | 75.1     |
| ENIGMA GL3 | 124985                 | 0979 | 85.8     |
| ENIGMA DX9 | 109373                 | 1011 | 72.6     |

#### Vertex format peer caching
Caching the native peer (in this case the actual vertex declaration) gave another huge performance boost to D3D9. I didn't want to mess up D3D11 so I did not touch it. I will come back to D3D11 later on when I have projection functions working and benchmark how to optimize it. OpenGL1 uses fixed function pipeline and client state so there's no real way to cache the vertex format at all. OpenGL3 is still using a single default VAO and sort of doing what GL1 does, so I didn't want to mess with it yet, but I did get a nice boost by using the cached lookup function for an attribute location.

| Branch     | Model create time (us) | FPS  | RAM (MB) | VRAM (MB) |
|------------|------------------------|------|----------|-----------|
| ENIGMA GL1 | 093785                 | 1406 | 75.1     | 50        |
| ENIGMA GL3 | 109386                 | 1220 | 86.1     | 72        |
| ENIGMA DX9 | 109397                 | 1483 | 69.4     | 38        |

#### Adjacent triangle strip combining using degenerates and buffer orphaning
I had reservations about doing this one, even though we did it before as well. Combining all primitive types together using an index buffer, which would also make fans work on mobile, is still an option. But in reality, there is nothing wrong in 2018 with using degenerate triangles to combine strips, even old graphics cards handle this really well.

During this phase, I discovered proper buffer orphaning and realized why TheExDeus had GL_STREAM_DRAW instead of GL_DYNAMIC_DRAW. Originally I did not properly understand the impact on performance this has when combined with buffer orphaning. By doing this first, I got a 50 fps boost in GL3, but then I realized instead of buffer orphaning we can just always call glBufferData because we never care about the previous contents of our buffer. By never calling `glBufferSubData` at all now, we avoid all synchronization overhead costs and I got an additional 50 fps boost in GL3 making the whole thing peak at 1607 max FPS. This is totally fine considering we don't have any intentions to support reading vertex buffers or updating subregions of them for now.

| Branch     | Model create time (us) | FPS  | RAM (MB) | VRAM (MB) |
|------------|------------------------|------|----------|-----------|
| ENIGMA GL1 | :small_red_triangle_down: 093785                 | :small_red_triangle_down: 1403 | :heavy_check_mark: 75.1     | :heavy_check_mark: 42        |
| ENIGMA GL3 | :small_red_triangle_down: 109386                 | :heavy_check_mark: 1607 | :small_red_triangle_down: 88.5     | :heavy_check_mark: 40        |
| ENIGMA DX9 | :small_red_triangle_down: 109366                 | :heavy_check_mark: 1711 | :heavy_check_mark: 69.6     | :heavy_minus_sign: 38        |